requirements.txt
backports.ssl-match-hostname==3.4.0.2
bottle==0.12.8
certifi==2015.4.28
colorama==0.3.3
dill==0.2.2
facebook-sdk==0.4.0
Flask==0.10.1
gensim==0.10.3
itsdangerous==0.24
Jinja2==2.7.3
MarkupSafe==0.23
nltk==3.0.0
nose==1.3.4
numpy==1.9.1
oauthlib==0.7.2
pathos==0.2a1.dev0
pox==0.2.1
ppft==1.6.4.5
processing===0.52-pathos
py==1.4.26
pymongo==2.8
pyre===0.8.2.0-pathos
pytest==2.6.4
requests==2.4.3
requests-oauthlib==0.4.1
scikit-learn==0.15.2
scipy==0.15.0
six==1.7.3
tldextract==1.5.1
tornado==4.2
tweepy==3.1.0
Werkzeug==0.9.6